manage installer based on custom fact
Prior to this commit the Gogs module ran exec's every execution to check
for the latest version.  This commit installs a custom fact so that you
can specify a version and compare this to the currently installed
version (stored in a fact).  This prevents resource executions when
there may not need to be one.